LayoutTests/accessibility/mac/text-input-marked-range-expected.txt

This test ensures that the input method marked range is available to accessibility clients.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !element.textInputMarkedRange is true
PASS element.textInputMarkedRange is '{0, 4}'
PASS !element.textInputMarkedRange is true
PASS element.textInputMarkedRange is '{1, 5}'
PASS successfullyParsed is true

TEST COMPLETE
